+++ src/Credfeto.Docker.HealthCheck.Http.Client/LoggingExtensions/HealthCheckClientLoggingExtensions.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Net;
 using Microsoft.Extensions.Logging;
 
 namespace Credfeto.Docker.HealthCheck.Http.Client.LoggingExtensions;
@@ -6,5 +7,12 @@ namespace Credfeto.Docker.HealthCheck.Http.Client.LoggingExtensions;
 internal static partial class HealthCheckClientLoggingExtensions
 {
     [LoggerMessage(EventId = 0, Level = LogLevel.Error, Message = "Healthcheck failed for {uri}")]
-    internal static partial void HealthCheckFailed(this ILogger logger, Uri uri, Exception exception);
+    internal static partial void HealthCheckFailed(this ILogger logger, string uri, Exception exception);
+
+    [LoggerMessage(
+        EventId = 1,
+        Level = LogLevel.Error,
+        Message = "Healthcheck returned unhealthy status {statusCode} for {uri}"
+    )]
+    internal static partial void HealthCheckUnhealthy(this ILogger logger, HttpStatusCode statusCode, string uri);
 }